At the end of auth, Server redirect directly to the embedded app URL
What:
Previously we redirected to the app index.  AppBridge would initialize on the frontend and then redirect to the embedded app URL. Now instead of initializing the frontend, only to incur a full page reload we do a server side redirect.

Why:
In our testing this is faster (8 seconds vs 10).  Additionally the UX is better because it cuts down on flashes of content that render and then quickly go away.